backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["momonong <morrischen2.9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.*"
fastapi = {extras = ["all"], version = "^0.103.1"}
pillow = "^10.0.0"
pandas = "^2.1.0"
uvicorn = "^0.24.0.post1"
openai = "^1.3.5"
python-dotenv = "^1.0.0"
streamlit = "^1.28.2"
langchain = "^0.0.340"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
